Move prior-turn sync persist out of export_precedence into exports.
Ensure side effects belong with catalog orchestration, keeping precedence
classification and payload resolution free of persistence I/O.
